Add glbc cluster
Update the local-setup script to include the creation of a separate kind
cluster to act as the private glbc control plane cluster that will have
cert manager installed on it.

Update the glbc controller to accept a `glbc-kubeconfig` option that
expects the kubeconfig of the glbc cluster, and adds an informer to
catch certificate events from that cluster.